requirements.txt
python-telegram-bot==20.7
openai==1.3.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
cryptography==42.0.5
